@@ -3,7 +3,6 @@ package com.example.mc_masterchemistry.UI.MejorPuntuacion;
 import android.content.Context;
 import android.os.Bundle;
 
-import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -15,16 +14,14 @@ import android.view.ViewGroup;
 
 import com.example.mc_masterchemistry.R;
 import com.example.mc_masterchemistry.db.Entities.Users;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.Query;
-import com.google.firebase.firestore.QuerySnapshot;
-import com.google.firebase.firestore.auth.User;
+
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 
 public class MejorPuntuacionFragment extends Fragment {
@@ -38,21 +35,10 @@ public class MejorPuntuacionFragment extends Fragment {
     public MejorPuntuacionFragment() {
     }
 
-
-    public static MejorPuntuacionFragment newInstance(int columnCount) {
-        MejorPuntuacionFragment fragment = new MejorPuntuacionFragment();
-        Bundle args = new Bundle();
-        args.putInt(ARG_COLUMN_COUNT, columnCount);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
         firestore=FirebaseFirestore.getInstance();
-
         if (getArguments() != null) {
             mColumnCount = getArguments().getInt(ARG_COLUMN_COUNT);
         }
@@ -62,7 +48,6 @@ public class MejorPuntuacionFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_mejor_puntuacion_list, container, false);
-
         // Set the adapter
         if (view instanceof RecyclerView) {
             Context context = view.getContext();
@@ -75,22 +60,16 @@ public class MejorPuntuacionFragment extends Fragment {
             firestore.collection("users")
                     .orderBy("gemas", Query.Direction.DESCENDING)
                     .get()
-                    .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-                        @Override
-                        public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                            alluser=new ArrayList<>();
-                            for (DocumentSnapshot document: task.getResult()){
-                                Users userItem = document.toObject(Users.class);
-                                alluser.add(userItem);
-                                adapter=new MyItemRecyclerViewAdapter(alluser);
-                                recyclerView.setAdapter(adapter);
-                            }
-
+                    .addOnCompleteListener(task -> {
+                        alluser=new ArrayList<>();
+                        for (DocumentSnapshot document: Objects.requireNonNull(task.getResult())){
+                            Users userItem = document.toObject(Users.class);
+                            alluser.add(userItem);
+                            adapter=new MyItemRecyclerViewAdapter(alluser);
+                            recyclerView.setAdapter(adapter);
                         }
                     });
-
         }
-
         return view;
     }
 
