@@ -9,16 +9,14 @@ import com.example.mc_masterchemistry.R;
 
 public class ExplicacionCompuestos extends AppCompatActivity {
 
-    private TextView titulo, informacion;
-
     private String nombre, datos;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_explicacion_compuestos);
-        titulo=findViewById(R.id.TV_titulo);
-        informacion=findViewById(R.id.TV_información);
+        TextView titulo = findViewById(R.id.TV_titulo);
+        TextView informacion = findViewById(R.id.TV_información);
         Intent();
         titulo.setText(nombre);
         informacion.setText(datos);
