@@ -1,6 +1,5 @@
 package com.example.mc_masterchemistry.UI.terciarios;
 
-import android.app.Application;
 import android.content.Context;
 
 
@@ -11,62 +10,60 @@ import java.util.List;
 
 
 public class Retos {
-   public static int numeroTipo=5;
-
-    private  int azar;
+   public static int numeroTipo = 5;
 
 
-
-    public Retos(Context ctx){
+    public Retos(){
 
 
     }
 
-    public String TipoReto(String retoSt){
-
+    public String TipoReto(){
+        String retoSt = "";
+        int azar;
         if(numeroTipo>4) {
 
             azar = (int) Math.floor(Math.random() * 5);
             if (tablero_terciarios.retoH==0&tablero_terciarios.retoH2O==0){
-                while(azar==3|azar==1){
+                while(azar ==3| azar ==1){
                     azar = (int) Math.floor(Math.random() * 5);
                 }
 
             }
             if (tablero_terciarios.retoH==1&tablero_terciarios.retoH2O==0){
-                while(azar==1) {
+                while(azar ==1) {
                     azar = (int) Math.floor(Math.random() * 5);
                 }
             }
             if (tablero_terciarios.retoH==0&tablero_terciarios.retoH2O==2){
-                while(azar==3) {
+                while(azar ==3) {
                     azar = (int) Math.floor(Math.random() * 5);
                 }
             }
 
 
             }else {
-           azar=numeroTipo;}
+           azar =numeroTipo;}
             switch (azar) {
                 case 0:
                     retoSt = " \n¡Comodín!, Puedes elegir un elemento de la lista con su NO. Combina los hidrogenos y oxígenos y escribelo directamente junto a ellos en la zona de crear."
                             +"\n\n"+ "      Fe  Co  Ni  Cu  Pd"+"\n\n"+"       Pt  Au  Hg  Tl  Sn"+"\n\n"+"      P  As  Sb  S  Se"+"\n\n"+"        Te  Cl  Br  I  At"+"\n";
-                    this.numeroTipo=0;
+                    numeroTipo=0;
                     break;
                 case 1:
                     retoSt = "Suma una molécula de agua a cualquier óxido de no metal o halogenuro de oxigeno, de los que  ya hayas creado. Escribelo en la zona de compuestos y fusiona. ";
-                    this.numeroTipo=1;
+                    numeroTipo=1;
                     break;
                 case 2:
                     retoSt = " Utiliza  los  cuatro elementos de oxígeno e hidrógeno (positivo o negativo) o pierdes 3 gemas. ";
-                    this.numeroTipo=2;
+                    numeroTipo=2;
                     break;
                 case 3:
                     retoSt = "Sustituye un H+ que este combinado con un No Metal por Au+, de los que ya hayas creado. Escribelo en la zona de compuestos y fusiona.";
-                    this.numeroTipo=3;
+                   numeroTipo=3;
                     break;
                 case 4:
-                    this.numeroTipo=4;
+                    numeroTipo=4;
                     retoSt = "Duplica cualquier carta de tus elementos para formar un nuevo compuesto";
 
                     break;
