@@ -47,7 +47,7 @@ public abstract class TiposRoomDatabase extends RoomDatabase {
 
 
 
-    private static RoomDatabase.Callback llamada = new RoomDatabase.Callback() {
+    private static final RoomDatabase.Callback llamada = new RoomDatabase.Callback() {
         @Override
         public void onCreate(@NonNull SupportSQLiteDatabase db) {
 
