@@ -13,14 +13,13 @@ import java.util.List;
 
 public class tiposViewModel extends AndroidViewModel {
 
-    private TiposRepository mRepository;
-    private LiveData<List<InorganicaEntity>> allinorganica;
+    private final LiveData<List<InorganicaEntity>> allinorganica;
 
 
     public tiposViewModel(@NonNull Application application) {
         super(application);
-        mRepository = new TiposRepository(application);
-        allinorganica=mRepository.getAll();
+        TiposRepository mRepository = new TiposRepository(application);
+        allinorganica= mRepository.getAll();
     }
 
     public LiveData<List<InorganicaEntity>> getAllElementos(){
